+++ Solutions/Day06/SolverDay6.cs
@@ -8,14 +8,14 @@ public class SolverDay6 : Solver
         _inputService = inputService;
     }
 
-    public override int SolvePartOne()
+    public override BigInteger SolvePartOne()
     {
         var input = GetInput();
 
         return 0;
     }
 
-    public override int SolvePartTwo()
+    public override BigInteger SolvePartTwo()
     {
         var input = GetInput();
         return 0;
